fix(locale): NewLocale cross platform
